Rename live-resolver package
