+++ packages/plugin-bower-manifest/package.json
@@ -10,7 +10,7 @@
     "@octolinker/helper-version": "1.0.0",
     "@octolinker/plugin-javascript": "1.0.0",
     "@octolinker/helper-regex-builder": "1.0.0",
-    "@octolinker/resolver-live-resolver-query": "1.0.0",
+    "@octolinker/resolver-live-query": "1.0.0",
     "@octolinker/resolver-git-url": "1.0.0",
     "@octolinker/resolver-github-shorthand": "1.0.0",
     "@octolinker/helper-insert-link": "1.0.0"
