+++ packages/plugin-composer-manifest/package.json
@@ -8,7 +8,7 @@
   "dependencies": {
     "@octolinker/helper-process-json": "1.0.0",
     "@octolinker/helper-regex-builder": "1.0.0",
-    "@octolinker/resolver-live-resolver-query": "1.0.0",
+    "@octolinker/resolver-live-query": "1.0.0",
     "@octolinker/helper-insert-link": "1.0.0"
   }
 }
