+++ packages/plugin-gemfile-manifest/package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "main": "./index.js",
   "dependencies": {
-    "@octolinker/resolver-live-resolver-query": "1.0.0",
+    "@octolinker/resolver-live-query": "1.0.0",
     "@octolinker/helper-grammar-regex-collection": "1.0.0"
   }
 }
