+++ packages/plugin-javascript/index.js
@@ -5,7 +5,7 @@ import {
   IMPORT,
   EXPORT,
 } from '@octolinker/helper-grammar-regex-collection';
-import liveResolverQuery from '@octolinker/resolver-live-resolver-query';
+import liveResolverQuery from '@octolinker/resolver-live-query';
 import builtinsDocs from './builtins-docs.js';
 
 function getTopModuleName(target) {
