+++ packages/plugin-rust/index.js
@@ -1,5 +1,5 @@
 import { RUST_CRATE } from '@octolinker/helper-grammar-regex-collection';
-import liveResolverQuery from '@octolinker/resolver-live-resolver-query';
+import liveResolverQuery from '@octolinker/resolver-live-query';
 
 export default {
   name: 'Rust',
