+++ packages/resolver-live-query/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "@octolinker/resolver-live-query",
+  "version": "1.0.0",
+  "description": "",
+  "repository": "https://github.com/octolinker/octolinker/tree/master/packages/resolver-live-query",
+  "license": "MIT",
+  "main": "./index.js"
+}
